+++ src/test/java/gov/usgs/wma/mlrlegacy/db/BaseIT.java
@@ -113,7 +113,13 @@ public abstract class BaseIT {
 	}
 
 	public String getInputJson(String file) throws IOException {
-		return replaceKey(getCompareFile("testData/", file));
+		createdDate = DEFAULT_CREATED_DATE_S;
+		createdBy = DEFAULT_CREATED_BY;
+		updatedDate = DEFAULT_UPDATED_DATE_S;
+		updatedBy = DEFAULT_UPDATED_BY;
+		return replaceKey(getCompareFile("testData/", file))
+				.replace("[createdDate]", createdDate).replace("[createdBy]", createdBy)
+				.replace("[updatedDate]", updatedDate).replace("[updatedBy]", updatedBy);
 	}
 
 	public String getExpectedReadJson(String file) throws IOException {