+++ src/test/java/gov/usgs/wma/mlrlegacy/db/ControllerCUIT.java
@@ -28,7 +28,7 @@ public class ControllerCUIT extends BaseControllerIT {
 	public void createSparseMonitoringLocation() throws Exception {
 		agency = UPDATED_AGENCY_CODE;
 		siteNbr = UPDATED_SITE_NUMBER;
-		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("sparseMonitoringLocation.json"), getAuthorizedHeaders());
+		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("createSparseMonitoringLocation.json"), getAuthorizedHeaders());
 
 		ResponseEntity<String> responseEntity = restTemplate.postForEntity("/monitoringLocations", entity, String.class);
 
@@ -52,7 +52,10 @@ public class ControllerCUIT extends BaseControllerIT {
 	public void createFullMonitoringLocation() throws Exception {
 		agency = UPDATED_AGENCY_CODE;
 		siteNbr = UPDATED_SITE_NUMBER;
-		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("fullMonitoringLocation.json"), getAuthorizedHeaders());
+		createdBy = DEFAULT_CREATED_BY;
+		updatedBy = DEFAULT_UPDATED_BY;
+		
+		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("createFullMonitoringLocation.json"), getAuthorizedHeaders());
 
 		ResponseEntity<String> responseEntity = restTemplate.postForEntity("/monitoringLocations", entity, String.class);
 
@@ -67,14 +70,14 @@ public class ControllerCUIT extends BaseControllerIT {
 
 	@Test
 	public void createNoToken() throws Exception {
-		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("sparseMonitoringLocation.json"), getHeaders());
+		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("createSparseMonitoringLocation.json"), getHeaders());
 		ResponseEntity<String> responseEntity = restTemplate.postForEntity("/monitoringLocations", entity, String.class);
 		assertEquals(401, responseEntity.getStatusCodeValue());
 	}
 
 	@Test
 	public void createUnauthorized() throws Exception {
-		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("sparseMonitoringLocation.json"), getUnauthorizedHeaders());
+		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("createSparseMonitoringLocation.json"), getUnauthorizedHeaders());
 		ResponseEntity<String> responseEntity = restTemplate.postForEntity("/monitoringLocations", entity, String.class);
 		assertEquals(403, responseEntity.getStatusCodeValue());
 	}
@@ -91,7 +94,7 @@ public class ControllerCUIT extends BaseControllerIT {
 		id = String.valueOf(ONE_MILLION);
 		agency = UPDATED_AGENCY_CODE;
 		siteNbr = UPDATED_SITE_NUMBER;
-		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("sparseMonitoringLocation.json"), getAuthorizedHeaders());
+		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("updateSparseMonitoringLocation.json"), getAuthorizedHeaders());
 
 		ResponseEntity<String> responseEntity = restTemplate.exchange("/monitoringLocations/" + id, HttpMethod.PUT, entity, String.class);
 
@@ -115,7 +118,9 @@ public class ControllerCUIT extends BaseControllerIT {
 		id = String.valueOf(ONE_MILLION);
 		agency = UPDATED_AGENCY_CODE;
 		siteNbr = UPDATED_SITE_NUMBER;
-		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("fullMonitoringLocation.json"), getAuthorizedHeaders());
+		createdBy = UPDATED_CREATED_BY;
+		updatedBy = UPDATED_MODIFIED_BY;
+		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("updateFullMonitoringLocation.json"), getAuthorizedHeaders());
 
 		ResponseEntity<String> responseEntity = restTemplate.exchange("/monitoringLocations/" + id, HttpMethod.PUT, entity, String.class);
 
@@ -133,11 +138,15 @@ public class ControllerCUIT extends BaseControllerIT {
 			table="legacy_location",
 			query=QUERY_ALL_TO_SECOND,
 			value="classpath:/testResult/oneSparsePatchDb/legacy_location.csv",assertionMode=DatabaseAssertionMode.NON_STRICT_UNORDERED,
-			modifiers={KeyModifier.class,CreatedModifier.class,UpdatedModifier.class}
+			modifiers={KeyModifier.class,CreatedModifier.class, UpdatedModifier.class}
 			)
 	public void updateMonitoringLocation_notFound() throws Exception {
 		id = String.valueOf(ONE_MILLION);
-		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("fullMonitoringLocation.json"), getAuthorizedHeaders());
+		agency = UPDATED_AGENCY_CODE;
+		siteNbr = UPDATED_SITE_NUMBER;
+		createdBy = UPDATED_CREATED_BY;
+		updatedBy = UPDATED_MODIFIED_BY;
+		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("updateFullMonitoringLocation.json"), getAuthorizedHeaders());
 
 		ResponseEntity<String> responseEntity = restTemplate.exchange("/monitoringLocations/1", HttpMethod.PUT, entity, String.class);
 
@@ -147,14 +156,14 @@ public class ControllerCUIT extends BaseControllerIT {
 
 	@Test
 	public void updateNoToken() throws Exception {
-		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("sparseMonitoringLocation.json"), getHeaders());
+		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("updateSparseMonitoringLocation.json"), getHeaders());
 		ResponseEntity<String> responseEntity = restTemplate.exchange("/monitoringLocations/" + id, HttpMethod.PUT, entity, String.class);
 		assertEquals(401, responseEntity.getStatusCodeValue());
 	}
 
 	@Test
 	public void updateUnauthorized() throws Exception {
-		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("sparseMonitoringLocation.json"), getUnauthorizedHeaders());
+		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("updateSparseMonitoringLocation.json"), getUnauthorizedHeaders());
 		ResponseEntity<String> responseEntity = restTemplate.exchange("/monitoringLocations/" + id, HttpMethod.PUT, entity, String.class);
 		assertEquals(403, responseEntity.getStatusCodeValue());
 	}
@@ -169,7 +178,7 @@ public class ControllerCUIT extends BaseControllerIT {
 			)
 	public void patchSparseMonitoringLocation() throws Exception {
 		id = String.valueOf(ONE_MILLION);
-		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("sparseMonitoringLocation.json"), getAuthorizedHeaders());
+		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("updateSparseMonitoringLocation.json"), getAuthorizedHeaders());
 
 		ResponseEntity<String> responseEntity = restTemplate.exchange("/monitoringLocations?_method=patch", HttpMethod.POST, entity, String.class);
 
@@ -191,7 +200,7 @@ public class ControllerCUIT extends BaseControllerIT {
 			)
 	public void patchFullMonitoringLocation() throws Exception {
 		id = String.valueOf(ONE_MILLION);
-		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("fullMonitoringLocation.json"), getAuthorizedHeaders());
+		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("updateFullMonitoringLocation.json"), getAuthorizedHeaders());
 
 		ResponseEntity<String> responseEntity = restTemplate.exchange("/monitoringLocations?_method=patch", HttpMethod.POST, entity, String.class);
 
@@ -267,16 +276,34 @@ public class ControllerCUIT extends BaseControllerIT {
 
 	@Test
 	public void patchNoToken() throws Exception {
-		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("sparseMonitoringLocation.json"), getHeaders());
+		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("updateSparseMonitoringLocation.json"), getHeaders());
 		ResponseEntity<String> responseEntity = restTemplate.exchange("/monitoringLocations?_method=patch", HttpMethod.POST, entity, String.class);
 		assertEquals(401, responseEntity.getStatusCodeValue());
 	}
 
 	@Test
 	public void patchUnauthorized() throws Exception {
-		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("sparseMonitoringLocation.json"), getUnauthorizedHeaders());
+		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("updateSparseMonitoringLocation.json"), getUnauthorizedHeaders());
 		ResponseEntity<String> responseEntity = restTemplate.exchange("/monitoringLocations?_method=patch", HttpMethod.POST, entity, String.class);
 		assertEquals(403, responseEntity.getStatusCodeValue());
+	}
+	
+	@Test
+	@DatabaseSetup("classpath:/testData/setupOneSparse/")
+	@ExpectedDatabase(
+			table="legacy_location",
+			query=QUERY_ALL_TO_SECOND,
+			value="classpath:/testResult/oneSparsePatchDb/legacy_location.csv",assertionMode=DatabaseAssertionMode.NON_STRICT_UNORDERED,
+			modifiers={KeyModifier.class,CreatedModifier.class,UpdatedModifier.class}
+			)
+	public void addDuplicateMonitoringLocation() throws Exception {
+		id = String.valueOf(ONE_MILLION);
+		agency = DEFAULT_AGENCY_CODE;
+		siteNbr = DEFAULT_SITE_NUMBER;
+		HttpEntity<String> entity = new HttpEntity<String>(getInputJson("createSparseMonitoringLocation.json"), getAuthorizedHeaders());
+
+		ResponseEntity<String> responseEntity = restTemplate.postForEntity("/monitoringLocations", entity, String.class);
+		assertEquals(406, responseEntity.getStatusCodeValue());
 	}
 
 }